+++ src/dexterity/localrolesfield/adapter.py
@@ -3,9 +3,11 @@
 from borg.localrole.interfaces import ILocalRoleProvider
 from dexterity.localroles.adapter import LocalRoleAdapter
 from dexterity.localrolesfield.field import LocalRolesField
+from dexterity.localrolesfield.interfaces import IBaseLocalRoleField
+from zope.interface import implements
+from plone.behavior.interfaces import IBehavior
 from plone.dexterity.interfaces import IDexterityFTI
 from zope.component import getUtility
-from zope.interface import implements
 
 
 class LocalRoleFieldAdapter(LocalRoleAdapter):
@@ -57,17 +59,25 @@ class LocalRoleFieldAdapter(LocalRoleAdapter):
         current context"""
         fti_schema = self.fti.lookupSchema()
         fields = [n for n, f in fti_schema.namesAndDescriptions(all=True)
-                  if isinstance(f, LocalRolesField)]
+                  if IBaseLocalRoleField.providedBy(f)]
 
         # also lookup behaviors
         for behavior_id in self.fti.behaviors:
             behavior = getUtility(IBehavior, behavior_id).interface
             fields.extend(
                 [n for n, f in behavior.namesAndDescriptions(all=True)
-                 if isinstance(f, LocalRolesField)])
+                 if IBaseLocalRoleField.providedBy(f)])
 
-        return [(f, v) for f in fields
-                for v in getattr(self.context, f) or []]
+        field_and_values = []
+        for field in fields:
+            values = getattr(self.context, field) or []
+            if not isinstance(values, list):
+                values = [values]
+
+            for value in values:
+                field_and_values.append((field, value))
+
+        return field_and_values
 
     @property
     def fti(self):