+++ src/dexterity/localrolesfield/interfaces.py
@@ -1,12 +1,24 @@
 # encoding: utf-8
+from zope.interface import Interface
+from zope.schema.interfaces import IChoice, IList
 
 from dexterity.localroles.interfaces import IDexterityLocalRoles
-from zope.schema.interfaces import IList
 
 
 class IDexterityLocalRolesField(IDexterityLocalRoles):
     """Specific layer for the package"""
 
 
-class ILocalRolesField(IList):
-    """Local roles field"""
+class IBaseLocalRoleField(Interface):
+
+    """Base interface for local role field."""
+
+
+class ILocalRolesField(IList, IBaseLocalRoleField):
+
+    """List local roles field."""
+
+
+class ILocalRoleField(IChoice, IBaseLocalRoleField):
+
+    """Choice local roles field."""