+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.2 (unreleased)
 ----------------
 
+- Add single value local role field.
+  [cedricmessiant]
+
 - Also lookup behaviors when searching for local role fields.
   [cedricmessiant]